requirements.txt
flaskflask
spacy
gunicorn
